Refactor `sim` package in prep for expanding the implementation
Break up large go files into separate ones that capture minimal
functionality. Encapsulate simulated signing related implementations
into their own package, called `signing`.

Rename inconsistent struct/constructor names, and resolve minor lint
issues.

Note, the changes here do not make any net functional difference.
Instead, they aim to reduce the LOC changed in successive PRs where
simulation will be expanded to accommodate more complex scenarios,
including more control over adversary message propagation, latency and
message loss modeling.
