+++ sim/signing/signing.go
@@ -1,0 +1,9 @@
+package signing
+
+import "github.com/filecoin-project/go-f3/gpbft"
+
+type Backend interface {
+	gpbft.Signer
+	gpbft.Verifier
+	GenerateKey() (gpbft.PubKey, any)
+}
